+++ cli/python/base_release/tests/test_engine_structure.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+from base_release import engine
+
+
+def test_release_readiness_helpers_are_split_from_engine() -> None:
+    engine_path = Path(engine.__file__)
+    readiness_path = engine_path.with_name("release_readiness.py")
+
+    assert readiness_path.exists()
+    assert "def release_findings" in readiness_path.read_text(encoding="utf-8")
+    assert "def release_findings" not in engine_path.read_text(encoding="utf-8")
+
+
+def test_release_publish_helpers_are_split_from_engine() -> None:
+    engine_path = Path(engine.__file__)
+    publish_path = engine_path.with_name("release_publish.py")
+
+    assert publish_path.exists()
+    assert "def run_release_step" in publish_path.read_text(encoding="utf-8")
+    assert "def run_release_step" not in engine_path.read_text(encoding="utf-8")
